pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = [
    {name = "Dina Khomenko", email = "dinamastercs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"  # Updated to match the dependencies section

dependencies = [
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "faker (>=37.0.2,<38.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.0"}

[tool.poetry.dependencies]
python = ">=3.9,<4.0"  # Matches the requires-python constraint
psycopg2-binary = "^2.9.10"